feat(InvestmentPerformanceScreen): Aggregate total portfolio performance
Display a combined view of all investments on the performance screen.

This change introduces the ability to select a new '__all__' option in the investment selector. When this option is chosen, the component aggregates the historical data from all individual investments to show the total market value and book value of the entire portfolio over time. This provides users with a holistic view of their investment growth.